+++ crates/polars-plan/src/dsl/functions/correlation.rs
@@ -1,7 +1,7 @@
 use super::*;
 
 /// Compute the covariance between two columns.
-pub fn cov(a: Expr, b: Expr, ddof: u8, min_periods: u8) -> Expr {
+pub fn cov(a: Expr, b: Expr, ddof: u8, min_periods: usize) -> Expr {
     let input = vec![a, b];
     let function = FunctionExpr::Correlation {
         method: CorrelationMethod::Covariance,
@@ -25,7 +25,7 @@ pub fn cov(a: Expr, b: Expr, ddof: u8, min_periods: u8) -> Expr {
 /// # Arguments
 /// * ddof
 ///     Delta degrees of freedom
-pub fn pearson_corr(a: Expr, b: Expr, ddof: u8, min_periods: u8) -> Expr {
+pub fn pearson_corr(a: Expr, b: Expr, ddof: u8, min_periods: usize) -> Expr {
     let input = vec![a, b];
     let function = FunctionExpr::Correlation {
         method: CorrelationMethod::Pearson,
@@ -54,7 +54,7 @@ pub fn pearson_corr(a: Expr, b: Expr, ddof: u8, min_periods: u8) -> Expr {
 ///     If to `false` then `NaN` are regarded as larger than any finite number
 ///     and thus lead to the highest rank.
 #[cfg(all(feature = "rank", feature = "propagate_nans"))]
-pub fn spearman_rank_corr(a: Expr, b: Expr, ddof: u8, propagate_nans: bool, min_periods: u8) -> Expr {
+pub fn spearman_rank_corr(a: Expr, b: Expr, ddof: u8, propagate_nans: bool, min_periods: usize) -> Expr {
     let input = vec![a, b];
     let function = FunctionExpr::Correlation {
         method: CorrelationMethod::SpearmanRank(propagate_nans),
